dns-resolver & EnvelopeAddress: explicitly check and deny ports in domain
This is prep work ahead of https://github.com/KumoCorp/kumomta/issues/352

This commit adds an explicit check for port numbers being passed in
via the domain when setting envelope addresses.

This is forbidden by the SMTP spec. This commit adds an explicit
integration test to confirm that we're reject them there.

For HTTP injection we don't have any other central point to
validate addresses, so this commit expands on the EnvelopeAddress::parse
method to add that centralized check.

dns-resolver's domain name validation is refactored into a domain
name classifier to help reuse that same logic outside the context
of making an actual DNS lookup.

MailExchanger::resolve is now also explicitly checking for and denying
lookups that include a port number.

This commit should have no practical functional impact; all of the
things denied by these checks are illegal in SMTP.