mailparsing: avoid flagging as LINE_TOO_LONG when rfc2047 header encoding
Despite taking care to wrap the header values that we produce, we're
still 2 characters over the limit of 78 that we consider to be
ideally conformant.

That will cause check_fix_conformance to rebuild the message content
to bring it within spec, but because the encoder is the same, it
won't actually satisfy the line length if you check-fix the fixed
message.

Let's just adjust the limit so that we can avoid that.