tsa-daemon: fix over zealous dedup of bounce/suspension data on websocket
We were deduping just by rule_hash, but each of these tables has
additional required fields as part of the primary key.

The result was that, for sites with a lot of bounces/suspensions
triggered by the same rules across a related set of sources,
the full set of bounces and suspensions would not be correctly
reported as part of a websocket push.