tsa-daemon: increase broadcast buffer to resolve channel lag error
We were using a fairly tight limit of 16 messages in the channel
that buffers the effects of changing bounces/suspensions from
any websocket-connected-clients.

A busy server could hit that limit fairly easily, resulting
in a `channel lagged by NUMBER` error that drops the websocket,
causing the client to need to reconnect and resync.

This commit resolves that by making the buffer a much more healthy size.