smtp_server: introduce ConcreteEsmtpListenerParams
The intent is that this will hold the merged, effective set
of configuration for a listener.

Right now this is a simple subset of the existing EsmtpListenerParams,
but in a future commit the latter will be extended and allow for
contextual parameters which will need to be compiled/merged
into the ConcreteEsmtpListenerParams struct.