queue/strategy: fix divergent behavior with skiplist strategy
This commit adds unit tests to validate the behavior when delaying
a message by a negative timestamp.

The expected behavior when that message is inserted into a scheduled
queue is for it to return it back to us instead of accepting it,
because it is immediately due.

This was true for 3 out of the 4 strategies; the skiplist strategy
allowed insertion but woke up the maintainer.  The net effect
of that should overall be equivalent, but it is technically
different.

This commit fixes it up so that it behaves the same as the
other strategies.