timeq: add timeq_pop_interval histogram
This tracks the time between calls to TimeQ::pop.
At the time of writing, this should be approx 3 seconds in a healthy system.
Occasional longer delays are OK, because the pop can catch up, but if
the system is continually lagging here, or lagging by a large amount,
then we can become tardy in dispatching messages from the scheduled
queue.