queue/maintainer.rs: allow processing popped batch async
This commit adds an undocumented and temporary knob for spawning
the reinsertion part of the singleton timer wheel in the interest
of reducing delays between wheel ticks.

From poking around at this, the default behavior is that the
singleton ticking wakes up on an arbitrary thread in the qmaint
pool.  That's normally desirable, but under heavy load that pool
is fairly saturated and that can lead to increased latency in
dispatching the ticks.

With this commit, when `kumo.set_timeq_spawn_reinsertion(true)`
is called during `pre_init`, we'll spawn a dedicated thread to
tick the singleton wheel, and each time it dispatches, it will
launch the spawned message batch into the qmaint pool.

This keeps the pop interval consistent.