systemd: read optional environment files
It is sometimes necessary to set environment variables
to adjust some early-startup options that need to be
applied before enough of the system is ready to read
lua policy files.

While it is possible to use `systemctl edit SERVICE`
and enter systemd-specific directives to configure that,
it is not the most convenient thing.

What this commit does is introduce 3 new optional files:

* /opt/kumomta/etc/kumo-common.env: sourced by kumod and tsa
* /opt/kumomta/etc/tsa-daemon.env: sourced by tsa
* /opt/kumomta/etc/kumod.env: sourced by kumod

The format of these files is `NAME=VALUE` on each line, as
per the `EnvironmentFile` docs in systemd.

One practical application for this is on an over-committed system
where the number of reported cores is larger than the number of
available cores.

For such a system you will need to configure the
`KUMO_AVAILABLE_PARALLELISM` environment variable for each kumo
service running on your system. With these environment files you
can do that simply by creating `/opt/kumomta/etc/kumo-common.env`
with a line like this inside:

```
KUMO_AVAILABLE_PARALLELISM=48
```