admin bounce: add caching+negative caching for queue insertion
For sites with large numbers of admin bounce entries, it is
desirable to negatively cache the lookup so that we can avoid
extraneous matching overhead.

We use a generation counter to determine if any rules have changed
and allow that to drive the invalidation of negative caching.

Otherwise, we cache the last lookup and, if the generation is the
same, then we cache the prior result and respect its expiration.