queue: add timeq_pop_tardiness, timeq_reinsert_tardiness histogram
These measures the difference between the due time and the current
time at the point where a message is popped from the singleton
wheel (timeq_pop_tardiness) and the time when we initiate a
call to reinsert the message into the ready queue
(timeq_reinsert_tardiness).

Since the wheel ticks every 3 seconds, we typically expect to
see the maximum value in timeq_pop_tardiness being ~3 seconds.

timeq_reinsert_tardiness should not generally be significantly
larger than this, but since there is some uncertainty, we need
this metric.