accounting: fixup blocking issue on flush
I noticed that sometimes CI will fail with an error mentioning
the accounting db and a timeout. It appears as though that is
a secondary issue masking the true problem.

Revise the logic to use spawn_blocking around the inherently
blocking sqlite query operation so that we can reveal the
true issue.